Expose most of heudiconv CLI interface
There are a couple of options that I didn't add here since they didn't
seem to fit well. Also there are a couple of additional ones for
things like git user details when using datalad.

There are a couple of known limitations, which I've created issues for
in this repo:

https://github.com/rh-impact/pl-heudiconv/issues/8
https://github.com/rh-impact/pl-heudiconv/issues/9